separating classes into suitable packages
what may be a problem is cyclical package class dependencies between the entities (needing the moveTo method from a level object) and the levels (needing the entity class to set up array).   look into setting up a newPosition flag that can be checked in the frame update loop, then remove level calls in the entities package;  have a set method for whether the position change suceeded or failed, whether the internal method should go back to the previous tile or not and clear the flag.  This will leave levels dependent on entities, and entities dependent on items.